запуск ботов сделан на разные файлы

+++ Telegramm/bot.py
@@ -1,15 +1,14 @@
 from aiogram import types, Dispatcher
 from functions.functions import get_discount_code, key_dict
-from create_bot import dp, bot
+from create_bot_tg import dp, bot
 from .keyboard import kb_client
+from .admin import admins, sallers
+import re
 
 
-# @dp.message_handler()
 async def check_client(message: types.Message):
-
     user_id = message.from_user.id  # ID пользователя который запустил хэндлер
-    # user_channel_status = await bot.get_chat_member(key_dict['tg_chanel'], user_id)
-    # if user_channel_status["status"] == "left":
+
     if False:
         await bot.send_message(message.chat.id, f'Подпишитесь на наш канал: {key_dict["url_tg"]}\n,'
                                                 f'чтобы получить купон и повторите попытку.')
@@ -25,20 +24,19 @@ async def check_client(message: types.Message):
         await message.delete()
 
 
-# @dp.message_handler(commands=['start', 'help'])
 async def command_start(message: types.Message):
-
-    try:
-        await bot.send_message(message.from_user.id, f"Если вы еще не получали скидочный купон, "
-                                                     f"нажмите 'Получить купон' или введите команду /coupon",
-                               reply_markup=kb_client)
-        await message.delete()
-    except:
-        await message.reply(f'Ошибка соединения...')
+    # if message.from_user.id not in admins and message.from_user.id not in sallers:
+    if not str(message.from_user.id) in admins and not str(message.from_user.id) in sallers:
+        try:
+            await bot.send_message(message.from_user.id, f"Если вы еще не получали скидочный купон, "
+                                                         f"нажмите 'Хочу купон' или введите сообщение 'Хочу купон'",
+                                   reply_markup=kb_client)
+            await message.delete()
+        except:
+            await message.reply(f'Ошибка соединения...')
 
 
 def register_handler_client(dp: Dispatcher):
 
     dp.register_message_handler(check_client, text='Хочу купон')
     dp.register_message_handler(command_start, state='*', content_types=['text'])
-
